requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
python-dotenv
cryptography
loguru
requests
python-multipart
pandas
ta
psycopg2-binary
asyncpg
